# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/image": {
            "post": {
                "description": "Uploads an image",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "operationId": "upload-image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Images to upload",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/image-optimization-api_pkg_server.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/api/image/list": {
            "get": {
                "description": "Retrieves a list of images",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "operationId": "list-images",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/image-optimization-api_internal_service_image.ListImageResponse"
                        }
                    }
                }
            }
        },
        "/api/image/{image_id}": {
            "get": {
                "description": "Retrieves an image by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "operationId": "get-image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "image_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Compression Quality (one of: 100, 75, 50, 25)",
                        "name": "quality",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/image-optimization-api_internal_service_image.GetImageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "image-optimization-api_internal_service_image.GetImageResponse": {
            "type": "object",
            "properties": {
                "image": {
                    "$ref": "#/definitions/image-optimization-api_internal_service_image.ImageInfo"
                }
            }
        },
        "image-optimization-api_internal_service_image.ImageInfo": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "image-optimization-api_internal_service_image.ListImageResponse": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/image-optimization-api_internal_service_image.ImageInfo"
                    }
                }
            }
        },
        "image-optimization-api_pkg_server.EmptyResponse": {
            "type": "object"
        }
    }
}